requirements.txt
llama-index
python-dotenv
openai
pyyaml